+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Added new standalone mode `hf_legicsim` (@uhei)
  - Changed `hf legic *` - now uses NG instead (@iceman1001)
  - Added `hf legic view` - view contents of LEGIC Prime dump files (@iceman1001)
  - Changed `hf mfu restore` - now takes bin/json as dump files (@iceman1001)